Move SkPaint ToSk to dl_sk_conversions for consistency.
